Refactored DEEPCore and DEEPQualtrics to remove tight coupling
Previously, DEEPCore would store a DEEPCore.platform variable and call methods inside of the platform. However, this doesn't seem like a good practice since it means that DEEPCore and the platform are tightly coupled. Instead, this refactor moves away from a situation where both the platform and DEEPCore push and pull, to a situation where the platform pushes to DEEPCore and also pulls from DEEPCore; DEEPCore itself does not invoke anything on the platform side. This means that we can decouple DEEPCore and the platform and allow the platform logic to operate independently, which decreases complexity.